Added City repository
