+++ app/src/main/java/com/chichkanov/yandex_weather/repository/CityRepository.java
@@ -1,0 +1,14 @@
+package com.chichkanov.yandex_weather.repository;
+
+import com.chichkanov.yandex_weather.model.places.CitySuggestion;
+
+import io.reactivex.Observable;
+
+
+public interface CityRepository {
+    Observable<CitySuggestion> getCitySuggestion(String cityName);
+
+    void setCurrentCity(String city);
+
+    String getCurrentCity();
+}
